+++ src/main/java/chess/service/dto/ScoreStatusDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.service.dto;
 
 import chess.domain.piece.Score;
 import chess.domain.piece.Team;
