+++ src/test/java/chess/service/ChessGameServiceTest.java
@@ -11,12 +11,12 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
-import chess.dto.BoardDto;
-import chess.dto.RankDto;
-import chess.dto.ScoreStatusDto;
 import chess.fixture.PositionFixtures;
 import chess.repository.PieceRepository;
 import chess.repository.TurnRepository;
+import chess.service.dto.BoardDto;
+import chess.service.dto.RankDto;
+import chess.service.dto.ScoreStatusDto;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
